FIX: icon links regression following core changes
In https://github.com/discourse/discourse/commit/34ffc0065ac17244109f7e6da0758cc85ec70bec
we added a `mouseDown` event hook to the search banner widget. That
widget hook triggers a rerender in core, which ends up stopping event
propagation and preventing the `click` event for these links.

This is a quick and dirty fix here, I think core should not rerender a
widget for the `mousedown` and `mouseup` hooks.