DEV: Lower Plugin API version (#35)
The search banner initializer was pointing to a recent version of the core Plugin API which could cause issues if a discourse instance was utilizing the `discourse-header-search` component but had not updated to a recent version of core. This PR brings the Plugin API version down to a version that will provide more wiggle room for people to run the theme component without issues on older versions of core.